test: pin the int32 page-offset overflow in _page_split_kernel
page_idx and sub come from tl.program_id(0), so both byte offsets in
_page_split_kernel are int32 products. Past 2**31 they wrap and Triton
sign-extends the result onto the 64-bit pointer, putting the access about
2 GiB below the buffer.

The two offsets wrap at different pool sizes, so each gets its own test:
the destination at dst page 57,359 (14,340 source pages), the source at
page_idx 14,340. Each test packs the axis it is not exercising, so only
one buffer spans 2**31 at a time and peak stays at 4.13 GiB. Both sit
above a canary region, so a regression trips an assertion instead of an
illegal access that would take down the CUDA context for the process.

Also carries the one-line int64 promotion from #34027 so the test is
green standalone.

+++ python/sglang/kernels/ops/attention/flash_mla_sm120.py
@@ -315,7 +315,8 @@ def _page_split_kernel(
     are copied; untouched pages are skipped so the kernel no longer rewrites the
     entire KV pool every decode step.
     """
-    pid = tl.program_id(0)
+    # int64 so the byte-offset products below cannot wrap past 2**31 on large pools
+    pid = tl.program_id(0).to(tl.int64)
     page_idx = pid // RATIO
     sub = pid % RATIO
